requirements.txt
langchain
sentence-transformers
faiss-cpu
PyMuPDF
streamlit
